moved AttributionEnseignantTranslator to common_domain
