+++ ddd/logic/encodage_des_notes/common_domain/service/i_attribution_enseignant.py
