+++ ddd/logic/encodage_des_notes/soumission/test/use_case/read/test_get_progression_generale_encodage_service.py
@@ -34,7 +34,7 @@ from ddd.logic.encodage_des_notes.tests.factory._note_etudiant import NoteManqua
 from ddd.logic.encodage_des_notes.tests.factory.feuille_de_notes import FeuilleDeNotesAvecUneSeuleNoteManquante, \
     FeuilleDeNotesAvecNotesEncodeesEtNotesManquantes, FeuilleDeNotesAvecNotesSoumises, \
     FeuilleDeNotesDateLimiteRemiseHierEtAujourdhui
-from infrastructure.encodage_de_notes.soumission.domain.service.in_memory.attribution_enseignant import \
+from infrastructure.encodage_de_notes.common_domain.service.in_memory.attribution_enseignant import \
     AttributionEnseignantTranslatorInMemory
 from infrastructure.encodage_de_notes.common_domain.service.in_memory.periode_encodage_notes import \
     PeriodeEncodageNotesTranslatorInMemory
