+++ ddd/logic/encodage_des_notes/soumission/use_case/read/get_progression_generale_encodage_service.py
@@ -24,7 +24,7 @@
 #
 ##############################################################################
 from ddd.logic.encodage_des_notes.soumission.commands import GetProgressionGeneraleCommand
-from ddd.logic.encodage_des_notes.soumission.domain.service.i_attribution_enseignant import \
+from ddd.logic.encodage_des_notes.common_domain.service.i_attribution_enseignant import \
     IAttributionEnseignantTranslator
 from ddd.logic.encodage_des_notes.common_domain.service.i_periode_encodage_notes import \
     IPeriodeEncodageNotesTranslator
