+++ infrastructure/encodage_de_notes/common_domain/service/attribution_enseignant.py
@@ -30,7 +30,7 @@ from django.db.models.functions import Concat
 
 from attribution.models.attribution_class import AttributionClass
 from attribution.models.attribution_new import AttributionNew
-from ddd.logic.encodage_des_notes.soumission.domain.service.i_attribution_enseignant import \
+from ddd.logic.encodage_des_notes.common_domain.service.i_attribution_enseignant import \
     IAttributionEnseignantTranslator
 from ddd.logic.encodage_des_notes.soumission.dtos import AttributionEnseignantDTO
