+++ infrastructure/encodage_de_notes/common_domain/service/in_memory/attribution_enseignant.py
@@ -25,7 +25,7 @@
 ##############################################################################
 from typing import Set
 
-from ddd.logic.encodage_des_notes.soumission.domain.service.i_attribution_enseignant import \
+from ddd.logic.encodage_des_notes.common_domain.service.i_attribution_enseignant import \
     IAttributionEnseignantTranslator
 from ddd.logic.encodage_des_notes.soumission.dtos import AttributionEnseignantDTO
