+++ infrastructure/tests/encodage_des_notes/domain/service/test_attribution_enseignant.py
@@ -27,11 +27,10 @@ from unittest import skip
 
 from django.test import TestCase
 
-from attribution.models.attribution_class import AttributionClass
 from attribution.tests.factories.attribution_class import AttributionClassFactory
 from attribution.tests.factories.attribution_new import AttributionNewFactory
 from ddd.logic.encodage_des_notes.soumission.dtos import AttributionEnseignantDTO
-from infrastructure.encodage_de_notes.soumission.domain.service.attribution_enseignant import \
+from infrastructure.encodage_de_notes.common_domain.service.attribution_enseignant import \
     AttributionEnseignantTranslator
